- Les opérateurs +/- de point3 retourne maintenant un vector3 - ajout d'un constructeur a Matrix4 - modification à camera afin de pouvoir retourner les matrice vue/projection - modification a beaucoup de projets, afin de compiler suite aux changements précédents.

+++ trunk/projects/RealEdit/Controller.cpp
@@ -53,14 +53,14 @@ void Controller::createPlatonicSolid(PlatonicSolid::type iType)
   
   vector<RealEditPoint> vPoints;
   vector<RealEditPoint> vFace;
-  for (uint i = 0; i < ps.getNumberOfVertices(); ++i) 
+  for (uint i = 0; i < (uint)ps.getNumberOfVertices(); ++i) 
   {    
     RealEditPoint p(ps.getVertex( i ));
     m.addPoint (p);
     vPoints.push_back(p);
   }
   
-  for (uint i = 0; i < ps.getNumberOfPolygons(); ++i)
+  for (uint i = 0; i < (uint)ps.getNumberOfPolygons(); ++i)
   {
     vFace.clear();
     const Polygon& p = ps.getPolygon( i );
